move methods out of front_landview and into api_campaign
