+++ src/front_landview.h
@@ -138,18 +138,7 @@ void check_mouse_scroll(void);
 void update_velocity(void);
 /******************************************************************************/
 void update_ensigns_visibility(void);
-TbBool campaign_level_api_set_available(LevelNumber lvnum);
-TbBool campaign_level_api_set_unavailable(LevelNumber lvnum);
-TbBool campaign_level_api_is_enabled(LevelNumber lvnum);
-TbBool campaign_level_api_is_available(LevelNumber lvnum);
-void campaign_level_api_reset(void);
-void campaign_level_api_refresh(void);
-TbBool get_campaign_auto_advance_enabled(void);
-void campaign_level_api_set_auto_advance(TbBool enabled);
-TbBool campaign_bonus_level_api_is_in_campaign(LevelNumber bn_lvnum);
-TbBool campaign_bonus_level_api_set_available(LevelNumber bn_lvnum);
-TbBool campaign_bonus_level_api_set_unavailable(LevelNumber bn_lvnum);
-TbBool campaign_bonus_level_api_is_available(LevelNumber bn_lvnum);
+
 
 #ifdef __cplusplus
 }
